+++ src/NUglify.Tests/TestData/JS/Expected/ES2020/GlobalThis.js
