+++ src/NUglify.Tests/TestData/JS/Expected/ES2021/LogicalAndAssign.js
