+++ src/NUglify.Tests/TestData/JS/Expected/ES2021/LogicalNullishAssign.js
