+++ src/NUglify.Tests/TestData/JS/Expected/ES2021/LogicalOrAssign.js
