+++ src/NUglify.Tests/TestData/JS/Expected/ES2021/NumericSeparatorsBinary1.js
