+++ src/NUglify.Tests/TestData/JS/Expected/ES2021/NumericSeparatorsBinary2.js
