+++ src/NUglify.Tests/TestData/JS/Expected/ES2021/NumericSeparatorsDecimal1.js
