+++ src/NUglify.Tests/TestData/JS/Expected/ES2021/NumericSeparatorsDecimal2.js
