+++ src/NUglify.Tests/TestData/JS/Expected/ES2021/NumericSeparatorsDecimalBefore.js
