+++ src/NUglify.Tests/TestData/JS/Expected/ES2021/NumericSeparatorsExponent2.js
