+++ src/NUglify.Tests/TestData/JS/Expected/ES2021/NumericSeparatorsExponentBefore.js
