+++ src/NUglify.Tests/TestData/JS/Expected/ES2021/NumericSeparatorsHex1.js
