+++ src/NUglify.Tests/TestData/JS/Expected/ES2021/NumericSeparatorsHex2.js
