+++ src/NUglify.Tests/TestData/JS/Expected/ES2021/NumericSeparatorsInteger.js
