+++ src/NUglify.Tests/TestData/JS/Expected/ES2021/NumericSeparatorsNoAdjacent.js
