+++ src/NUglify.Tests/TestData/JS/Expected/ES2021/NumericSeparatorsNoTrailing.js
