+++ src/NUglify.Tests/TestData/JS/Expected/ES2021/NumericSeparatorsOctal1.js
