+++ src/NUglify.Tests/TestData/JS/Expected/ES2021/NumericSeparatorsOctal2.js
