+++ src/NUglify.Tests/TestData/JS/Input/ES2020/GlobalThis.js
