+++ src/NUglify.Tests/TestData/JS/Input/ES2021/NumericSeparatorsBinary2.js
