+++ src/NUglify.Tests/TestData/JS/Input/ES2021/NumericSeparatorsDecimal1.js
