+++ src/NUglify.Tests/TestData/JS/Input/ES2021/NumericSeparatorsDecimal2.js
