+++ src/NUglify.Tests/TestData/JS/Input/ES2021/NumericSeparatorsDecimalAfter.js
