+++ src/NUglify.Tests/TestData/JS/Input/ES2021/NumericSeparatorsDecimalBefore.js
