+++ src/NUglify.Tests/TestData/JS/Input/ES2021/NumericSeparatorsExponent2.js
