+++ src/NUglify.Tests/TestData/JS/Input/ES2021/NumericSeparatorsHex1.js
