+++ src/NUglify.Tests/TestData/JS/Input/ES2021/NumericSeparatorsHex2.js
