+++ src/NUglify.Tests/TestData/JS/Input/ES2021/NumericSeparatorsNoTrailing.js
