+++ src/NUglify.Tests/TestData/JS/Input/ES2021/NumericSeparatorsOctal1.js
