+++ src/NUglify.Tests/TestData/JS/Input/ES2021/NumericSeparatorsOctal2.js
